Importing package for CPAN module CPAN::Meta::YAML version 0.007 into devel/p5-CPAN-Meta-YAML as dependency of scheduled update of package for CPAN module CPAN::Meta in devel/p5-CPAN-Meta.
This module implements a subset of the YAML specification for use in
reading and writing CPAN metadata files like META.yml and MYMETA.yml.
It should not be used for any other general YAML parsing or generation
task.

NOTE: META.yml (and MYMETA.yml) files should be UTF-8 encoded. Users
are responsible for proper encoding and decoding. In particular, the
read and write methods do not support UTF-8 and should not be used.