Update to 1.1.1.
1.1.1:
This release builds on the DocBook/XML conversion of the XDMCP spec
included in the 1.1.0 release with cleanups & improvements to make
the documentation more useful and usable.

1.1.0:
This release features a major refactoring and cleanup of the code base,
and conversion of the XDMCP specification from troff to DocBook/XML,
along with the usual recent set of build configuration improvements and
janitorial cleanups.