Update to 2.3.1:
2.3.1:
This bug fix release corrects a regression in font emboldening introduced
in 2.3.0.   Another bug against font appearance in 2.3.0 has been reported
but not yet resolved - since no progress was being made on it, I decided
to stop holding back the release of the fix we have.

2.3.0:
Notable changes in this release include:

 - Subpixel LCD text rendering improvements

 - Dropped support for building against versions of freetype2, fontconfig &
   libXrender that pre-dated pkgconfig support.

 - Raised minimum required versions to freetype2 2.1.6 & fontconfig 2.5.92.

plus of course assorted bug fixes and build configuration improvements.

2.2.0:
This release removes the deprecated xft-config script now that pkg-config
is well-established as it's replacement.   It also updates the xft pkg-config
file to reduce unnecessary linking in users of libXft, but may cause issues if
used with versions of pkg-config older than the pkg-config 0.22 release from
2007.

It also includes the usual recent set of build configuration improvements and
janitorial cleanups.