Update to 1.3.2.
1.3.2:
This minor maintenance release mainly makes the man pages less sucky.
The build configuration was also made to suck a little less too.

1.3.1:
This release includes several fixes to the error handling paths
in the library, as well as the usual set of recent build configuration
improvements and janitorial cleanups.