net/coda: Back out DragonFly patches
Coda requires a kernel module to run and DragonFly had such a kernel at
one point.  However the kernel was removed six years ago and there is no
desire to restore it.  One of the main reasons against doing so is the
lack of IPv6 support in Coda.

The userland patches for coda are therefore unnecessary and coda has been
reverted to its previous state save for marking NOT-FOR-DRAGONFLY.

PKGREVISION was mispelled in the makefile, so fixing that typo bumps the
revision to nb2 (caught by pkglint).