Remove Asterisk 1.6.  This version series went end-of-line on April 21st, 2012.  It most likely has multiple security issues.  By this point, all users of this package should have migrated to comms/asterisk18 or comms/asterisk10 as this version has been marked as being deprecated for some time now.
Note that this directory is likely to re-appear in late 2017 when
Asterisk 16 comes out, assuming the current schedule is followed.
However that will be a vastly different version as Asterisk 11 is
only in the RC stage now (i.e. it will be five major versions after
the one that is expected to be released later this year).