Update py-xlrd to 0.8.0.
0.8.0:
This release finally lands the support for both .xls and .xlsx files.
Many thanks to John Machin for all his work on making this happen.
Opening of .xlsx files is seamless, just use xlrd as you did before and
it all should "just work".

xlrd 0.8.0 is also the first release that that targets Python 2.6 and
2.7, but no Python 3 just yet. Python 2.5 and below may work but are not
supported. If you need to use Python 2.5 or earlier, please stick to
xlrd 0.7.x.

0.7.9:
This release fixes an annoying merge bug on my part that resulted in a
"NameError: global name 'BYTES_X00' is not defined" error where opening
certain Excel files.

0.7.8:
This release features the following changes:

- Compatibility with Python 2.1 and 2.2 is restored.
- Fix for github issue #7: assertion error when reading file with
  xlwt-written bitmap. The assertion is now ignored and a warning logged
  if verbosity > 0.
- superfluous zero bytes at end of xls OBJECT records are now ignored.