Import icinga-base-1.8.4 as net/icinga-base.
Icinga is a program that will monitor hosts and services on your
network.  It has the ability to email or page you when a problem arises
and when it gets resolved.  Icinga is written in C and is designed to
run under Linux, although it should work under most other *NIX variants.
It can run either as a normal process or as a daemon, intermittently
running checks on various services that you specify.  The actual service
checks are performed by external "plugins" which return service
information to Icinga.  Several CGI programs are included with Icinga
in order to allow you to view the current service status, history, etc.
via a web browser.

This package provides the core Icinga infrastructure and web pages.

Icinga is a fork of Nagios (net/nagios-base) and Icinga uses
Nagios's plugins (net/nagios-plugins).