Update to 1.7.1: Only a single fix: including the XFixes header to get the typedef for PointerBarrier. Naturally, this adds a pkgconfig build-time dependency on XFixes.
Why? The header shipped with 1.7 typedef'd PointerBarrier. If you #include
both XI and Xfixes headers, you will end up with a duplicate typedef.

This is not an issue on gcc >= 4.6 since the two typedef's are the same.
On earlier versions this will trigger an error. gcc 4.6 -pedantic-errors
will trigger the same error.

Peter Hutterer (2):
      Require XFixes for PointerBarrier, remove duplicate typedef                                                                                                                    libXi 1.7.1