Update to 0.80.81.
Upstream did not provide NEWS.  But, this is ~4 years of assorted
bugfixes and minor improvements.  Most notably there is a fix that is
critical for stream startup in a timing-dependent case which seems to
hit reliably on netbsd-5/i386-xen, jabberd2.

Take MAINTAINERSHIP, as I appear to also be the upstream maintainer
now.